Refactoring shell scripts. Cleaned up conditionals.

+++ getweather.sh
@@ -16,40 +16,27 @@ echo "Path: ${_}"
 function parse_return {
   rrr=$(echo $1 | tr '\n' ' ' | sed 's/ $//g')
   r2=$(echo ${rrr} | grep -e "corrupt image" -e "improper image header" -e "no images defined" -e "unable to open image")
-  if [[ "${r2}" != "" ]]; then
-    echo "Error! image is corrupted?"
-    exit
-  fi
+  [ "${r2}" != "" ] && echo "Error! image is corrupted?" && exit 1
   return 0
 }
 
-if [ "${convert_binary}" == "convert not found" ]; then
-  echo "Error! -convert- command not found."
-  exit
-fi
-
-if [ "${convert_binary}" == "" ]; then
-  echo "Error! -convert- command not found -- missing binary?"
-  exit
-fi
+/usr/bin/which convert | grep "convert not found" && exit
+# echo "Found Imagemagick binary. Moving forward."
+convert_binary=`source ${HOME}/.bash_profile; /usr/bin/which convert`
 
 for file in bkg2
 do
-  if [[ -f "${dir}/${file}.gif" ]]; then
-    sleep 1
-  else
-    echo "${file}.gif is not present in ${dir}. Copying it in."
+  [ -s "${dir}/${file}.gif" ] || echo "${file}.gif is not present in ${dir}. Copying it in." && \
     cp ${sourcelib}/images/${file}.gif ${dir}/
-  fi
 done
 
-if [[ -f "${dir}/trim_legend.gif" ]]; then
+if [[ -s "${dir}/trim_legend.gif" ]]; then
   # echo "Found trim_legend.gif file -- looks good."
   sleep 0.1
   else
     legendfile=$(ls -1 ${dir}/ | grep -e "N0R_Legend" -i | tr '\n' ' ' | sed  's/ //g')
 
-    if [[ -f "${dir}/${legendfile}" ]]; then
+    if [[ -s "${dir}/${legendfile}" ]]; then
       echo "No trim_legend.gif file found, but can convert ${legendfile}..."
       echo "Converting legend to crop1.gif:"
       # echo "convert FWS_N0R_Legend_0.gif -crop 0x0+0+25 crop1.gif"
@@ -62,10 +49,7 @@ fi
 
 for file in weather wow-test
 do
-  if [[ -f "${dir}/${file}.gif" ]]; then
-    # echo "Removing ${dir}/${file}.gif"
-    rm ${dir}/${file}.gif
-  fi
+  [ -s "${dir}/${file}.gif" ] && rm ${dir}/${file}.gif
 done
 
 # Get the MD5 hash of the backup image (no time stamp):
@@ -75,25 +59,24 @@ sha0=$(shasum ${dir}/wow_00.gif | awk {'print $1'} | tr '\n' ' ' | sed 's/ $//g'
 result=$(${convert_binary} ${dir}/current_image.gif -background black -alpha remove ${dir}/current_image.gif)
 # echo "returned text from convert is: ${result}"
 parse_return "${result}"
-result=""
 
+result=""
 result=$(${convert_binary} -composite ${dir}/current_image.gif  ${dir}/current_warnings.gif ${dir}/weather.gif)
 # echo "returned text from convert is: ${result}"
 parse_return "${result}"
-result=""
 
+result=""
 result=$(${convert_binary} -composite ${dir}/weather.gif ${dir}/bkg2.gif  ${dir}/wow-test.gif)
 # echo "returned text from convert is: ${result}"
 parse_return "${result}"
-result=""
 
+result=""
 result=$(${convert_binary} -composite ${dir}/wow-test.gif ${dir}/trim_legend.gif ${dir}/wow.gif)
 # echo "returned text from convert is: ${result}"
 parse_return "${result}"
 
 # Copy the potentially new image (no time stamp) to a backup image:
 cp ${dir}/wow.gif ${dir}/wow_00.gif
-
 
 # Make an MD5 hash of the potentially new image:
 sha1=$(shasum ${dir}/wow.gif | awk {'print $1'} | tr '\n' ' ' | sed 's/ $//g')
@@ -103,10 +86,7 @@ ${convert_binary} -font Courier -pointsize 18 -strokewidth 0.5 -stroke white \
   -draw "fill white text 530,50 '$(${date_binary} +%H:%M)'" ${dir}/wow.gif ${dir}/wow.gif
 
 # Check the two hashes against each other (backup to possibly-new):
-if [[ "${sha0}" == "${sha1}" ]]; then
-  echo "Radar image has not changed."
-  exit
-fi
+[ "${sha0}" == "${sha1}" ] && echo "Radar image has not changed." && exit
 
 # Copy the new time-stamped image to frame 0:
 cp ${dir}/wow.gif ${dir}/wow_0.gif
@@ -117,7 +97,7 @@ animate=" -delay 100 -dispose Background "
 for i in $(seq ${imagecount} -1 0)
   do
     sum=`expr ${i} - 1`
-    if [[ -f "${dir}/wow_${sum}.gif" ]]; then
+    if [[ -s "${dir}/wow_${sum}.gif" ]]; then
       # echo "Converting image from ${sum} to ${i}"
       mv ${dir}/wow_${sum}.gif ${dir}/wow_${i}.gif
       animate="${animate} ${dir}/wow_${i}.gif"
